new movements

+++ csc121-creative-project/src/Flag.java
@@ -34,19 +34,19 @@ public class Flag {
 	}
 	
 	/** Gets the x of a flag's posn */
-	public int getX() {
+	public double getX() {
 		return this.pos.getX();
 	}
 	
 	/** Gets the y of a flag's posn */
-	public int getY() {
+	public double getY() {
 		return this.pos.getY();
 	}
 	
 	/** produces the image of the current game */
 	public PApplet draw(PApplet c) {
 		c.fill(0, 0, 255);
-		c.rect(this.getX(), this.getY(), 20, 30);
+		c.rect((int)this.getX(), (int)this.getY(), 20, 30);
 		return c;
 	}
 }
